==> notebooks/30_spark_sql.ipynb <==
# Initialisation de Spark

from pyspark.sql import SparkSession
spark = SparkSession \
    .builder \
    .appName("Exemples avec les formats de stockage") \
    .getOrCreate()
sc = spark.sparkContext

# Chargement des données
1. Créer un dataframe pour les prénoms

prenoms = spark.read.load("prenomsParDeptsEtAnnees.parquet")
prenoms.createOrReplaceTempView("prenoms")
prenoms.show()

1. Faire de même pour les départements

#TODO
depts = spark.read.load("depts.parquet")
depts.createOrReplaceTempView("depts")
depts.show()

# Interroger les données

* La documentation sur Spark SQL est disponible [ici](https://spark.apache.org/docs/latest/sql-programming-guide.html) (de même pour [DataFrame](https://spark.apache.org/docs/latest/api/python/reference/pyspark.sql/api/pyspark.sql.DataFrame.html))
* Pour chaque question, vous donnerez la réponse avec le DSL et en SQL

1. Donner, pour chaque prénom, son nombre d'occurences

## DSL

#TODO
from pyspark.sql import functions as F

# total des effectifs par prénom (DSL)
prenoms.groupBy("prenom") \
       .agg(F.sum("effectif").alias("total")) \
       .orderBy(F.desc("total")) \
       .show(20)

## SQL

#TODO
# total des effectifs par prénom (SQL)
spark.sql("""
    SELECT prenom, SUM(effectif) AS total
    FROM prenoms
    GROUP BY prenom
    ORDER BY total DESC
    LIMIT 20
""").show()

1. Donner le nombre total de naissances avec un prénom féminin

## DSL

#TODO
prenoms.filter(F.col("sexe") == "2") \
        .agg(F.sum("effectif").alias("total_feminines")) \
        .show()

## SQL

#TODO
spark.sql("""
    SELECT SUM(effectif)
    FROM prenoms
    WHERE sexe = "2"
""").show()

1. Donner l'effectif maximal et minimal par prénom

## DSL

#TODO
prenoms.groupBy("prenom") \
       .agg(F.max("effectif").alias("max(effectif)"), F.min("effectif").alias("min(effectif)")) \
       .show(20)

## SQL

#TODO
spark.sql("""
    SELECT prenom, Max(effectif), Min(effectif)
    FROM prenoms
    GROUP BY prenom
    LIMIT 20
""").show()

1. Donner, pour chaque nom de département, le nombre moyen de prénoms depuis l'année 2000

## DSL

#TODO
prenoms.join(depts, prenoms.dep == depts.dep) \
       .filter(F.col("annee") >= 2000) \
       .groupBy(F.col("ncc")) \
       .agg(F.avg("effectif").alias("avg_effectif")) \
       .orderBy(F.desc("avg_effectif")) \
       .show(20)

## SQL

#TODO
spark.sql("""
    SELECT ncc, AVG(effectif)
    FROM prenoms p join depts d ON p.dep = d.dep
    WHERE p.annee >= 2000
    GROUP BY ncc
    ORDER BY AVG(effectif) DESC
    LIMIT 20
""").show()
